Update analysis workflow (#19)
This PR updates the analytical workflow. It now runs gets the estimand
and ACE after applying each step of the workflow, then calculates the
deltas against the original ACE for all steps after the initial one.

It's set up to automatically run on all examples in the repo that have
data, so it's important now to properly get data in (e.g. in #17) that
are _fully_ documented (otherwise, these results can't be
contextualized)