Update y0 usage and add notebook utils (#25)
This update does the following:

1. Externalizes some functionality that has been re-implemented upstream
in `y0`.
   - `add_ci_undirected_edges` is now part of `y0`
- `find_all_nodes_in_causal_paths` now has a more performant
implementation in `y0`
- the actual call to the evans simplification functions from y0 is now
used, which has a much more clean interface.
2. Moves shared code from the case study notebooks into
`eliater.notebook_utils` such that it can be reused. Further, this code
automatically adds descriptive text as well as more context than was
previously available in the case study notebooks. There are still some
open questions about the workflow for interpreting real data - from now
on, it makes more sense to implement this in the Python package rather
than having long runaway notebooks with tons and tons of code
3. the `_adjustment_set` argument was added into the eliater linear
regression to allow for pre-caching inside the notebook workflow
4. update to modernize setup.cfg. This includes the updates based on
https://github.com/y0-causal-inference/y0/issues/218 and
https://github.com/y0-causal-inference/y0/pull/219